Re-arrange components and pages
