+++ components/client/login.js
@@ -13,29 +13,36 @@ import {
     Stack,
     Button,
 } from '@chakra-ui/core';
-import { getErrorMessage } from '../lib/auth/form';
-import Container from './container';
+import { getErrorMessage } from '../../lib/auth/form';
+import Container from '../container';
 
-const AuthenticateAsRootMutation = gql`
-    mutation AuthenticateAsRootMutation($password: String!) {
-        authenticateAsRoot(password: $password)
+const AuthenticateAsClientMutation = gql`
+    mutation AuthenticateAsClientMutation($email: String!, $password: String!) {
+        authenticateAsClient(email: $email, password: $password) {
+            id
+            name
+            email
+        }
     }
 `;
 
-const StaffLogin = () => {
+const ClientLogin = () => {
     const client = useApolloClient();
-    const [authenticateAsRoot] = useMutation(AuthenticateAsRootMutation);
+    const [authenticateAsClient] = useMutation(AuthenticateAsClientMutation);
     const { handleSubmit, errors, setError, register, formState } = useForm();
 
-    const onSubmit = async ({ password }) => {
+    const onSubmit = async ({ email, password }) => {
         try {
             await client.resetStore();
-            const { data } = await authenticateAsRoot({
-                variables: { password },
+            const { data } = await authenticateAsClient({
+                variables: {
+                    email,
+                    password,
+                },
             });
 
-            if (data.authenticateAsRoot) {
-                Router.reload(window.location.pathname);
+            if (data.authenticateAsClient) {
+                Router.push('/client');
             }
         } catch (error) {
             setError('password', { message: getErrorMessage(error) });
@@ -53,9 +60,17 @@ const StaffLogin = () => {
                 mt="10"
                 maxWidth="900px"
             >
-                <Heading as="h1">Staff Login</Heading>
+                <Heading as="h1">Client Login</Heading>
                 <form onSubmit={handleSubmit(onSubmit)}>
                     <FormControl isInvalid={errors.password} isRequired>
+                        <FormLabel htmlFor="email">Email</FormLabel>
+                        <Input
+                            name="email"
+                            type="email"
+                            placeholder="Enter client's email address"
+                            ref={register}
+                            mb={4}
+                        />
                         <FormLabel htmlFor="password">Password</FormLabel>
                         <InputGroup size="md">
                             <Input
@@ -90,4 +105,4 @@ const StaffLogin = () => {
     );
 };
 
-export default StaffLogin;
+export default ClientLogin;
