+++ pages/client/login.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import ClientLoginPage from '../../components/client/login';
+
+const Clients = () => {
+    return <ClientLoginPage />;
+};
+
+export default Clients;
