--- pages/login.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ClientLogin from '../components/client-login';
-
-const Clients = () => {
-    return <ClientLogin />;
-};
-
-export default Clients;
