+++ pages/staff/login.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import StaffLoginPage from '../../components/staff/login';
+
+const Staff = () => {
+    return <StaffLoginPage />;
+};
+
+export default Staff;
